server04/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
av==15.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
coloredlogs==15.0.1
config==0.5.1
ctranslate2==4.6.0
distro==1.9.0
faster-whisper==1.2.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.26
logger==1.4
mpmath==1.3.0
numpy==2.3.2
onnxruntime==1.22.1
openai==1.106.1
orjson==3.11.3
packaging==25.0
piper-tts==1.3.0
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
zstandard==0.24.0
